requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
asyncpg==0.30.0
django==5.2.7
djangorestframework==3.16.1
docstring-parser==0.17.0
jiter==0.11.0
mirascope==1.25.6
pydantic==2.12.0
pydantic-core==2.41.1
python-dotenv==1.1.1
sqlparse==0.5.3
typing-extensions==4.15.0
typing-inspection==0.4.2
requests==2.32.3
psycopg2-binary==2.9.10
google-generativeai